requirements.txt
torch>=1.9.0
transformers>=4.21.0
tweepy>=4.14.0
requests>=2.28.0
python-dotenv>=0.19.0
numpy>=1.21.0
flask>=2.0.0